requirements.txt
dagit
dagster
dagster-postgres
duckdb
mysql-connector-python
numpy
pandas
psycopg2-binary
pytest
pyarrow
SQLAlchemy



